rename to types.internal
