+++ packages/react-component-event-listener/src/EventListener.ts
@@ -4,7 +4,7 @@ import addEventListener from './lib/addEventListener'
 import removeEventListener from './lib/removeEventListener'
 import shouldUpdateListener from './lib/shouldUpdateListener'
 import { EventListenerProps } from './types'
-import { listenerPropTypes } from './types.private'
+import { listenerPropTypes } from './types.internal'
 
 class EventListener extends React.Component<EventListenerProps> {
   static displayName = 'EventListener'
