+++ packages/react-component-event-listener/src/StackableEventListener.ts
@@ -5,7 +5,7 @@ import removeEventListener from './lib/removeEventListener'
 import * as listenerRegistries from './lib/listenerRegistries'
 import shouldUpdateListener from './lib/shouldUpdateListener'
 import { EventListenerProps } from './types'
-import { listenerPropTypes } from './types.private'
+import { listenerPropTypes } from './types.internal'
 
 class StackableEventListener extends React.Component<EventListenerProps> {
   static displayName = 'StackableEventListener'
