+++ packages/react-component-event-listener/src/types.internal.ts
